blank/requirements.txt
# pip install -r requirements.txt

psycopg2
python-dotenv
